fix: handle no workspace selected
When there is no folder selected and the Trivy extension is activated,
it can create the result storage location. Make this a little more
defensive.

This also highlighted an issue when the extension is only activated
once, updated in this PR to handle workspace folder change so that it
will reactivate and update the result storage location

Resolves #101

Signed-off-by: Owen Rumney <owen.rumney@aquasec.com>