feat: Add support for SAST results
Runs using Aqua Platform can optionally include SAST results. This
option is on at a workspace scoped level by default and can be turned
on and off same as DOTNET etc.

When there are results for SAST, they will be included in the existing
Treeview and problems

Updates the phases of the running updates to be more descriptive and
more extendable

Update the envvars to use consistent approach