feat: add default skip dirs
Add a list of default skip directories. These directories are overridden
at a workspace level through the standard vscode settings.

The skipdirs can be added to or removed from, if there are any in the
list when the scan runs, it will add them using the --skip-dirs flag for
Trivy or the Aqua plugin